// src/charting_library.d.ts
export interface PineJS {
    Std: PineJSStd;
}